Revert "Fix incorrect casting to avoid precision loss (#1501)" (#1536)
Reverts tensorflow/tflite-micro#1501

@mansnils from the CMSIS-NN team and I just chatted. This PR and #1491 are actually introducing a difference between Lite and Micro. If we do need these changes they would have to be made in both the TFLM and the TfLite kernels.

Before going down that path (which would involve a reasonable amount of time and effort from everyone involved), we would like to better understand what the issue reported in #1492 is and if that can be reproduced by the CMSIS-NN team as well.

BUG=#1492

+++ tensorflow/lite/micro/kernels/svdf_common.cc
@@ -451,21 +451,19 @@ TfLiteStatus PrepareSvdf(TfLiteContext* context, TfLiteNode* node) {
 
     TF_LITE_ENSURE_TYPES_EQ(context, output->type, kTfLiteInt8);
 
-    const double effective_scale_1 =
-        static_cast<double>(input->params.scale) *
-        static_cast<double>(weights_feature->params.scale) /
-        static_cast<double>(activation_state->params.scale);
+    const double effective_scale_1 = static_cast<double>(
+        input->params.scale * weights_feature->params.scale /
+        activation_state->params.scale);
     const double effective_scale_2 =
-        static_cast<double>(activation_state->params.scale) *
-        static_cast<double>(weights_time->params.scale) /
-        static_cast<double>(output->params.scale);
+        static_cast<double>(activation_state->params.scale *
+                            weights_time->params.scale / output->params.scale);
 
     // TODO(b/162018098): Use TF_LITE_ENSURE_NEAR when it is ready.
     TF_LITE_ENSURE(
         context,
         std::abs(static_cast<double>(bias->params.scale) -
-                 (static_cast<double>(activation_state->params.scale) *
-                  static_cast<double>(weights_time->params.scale))) < 1e-5);
+                 static_cast<double>(activation_state->params.scale *
+                                     weights_time->params.scale)) < 1e-5);
 
     QuantizeMultiplier(effective_scale_1, &(data->effective_scale_1_a),
                        &(data->effective_scale_1_b));
